tests/wpt/web-platform-tests/workers/constructors/SharedWorker/connect-event.html

<!doctype html>
<title>connect event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new SharedWorker('connect-event.js');
  worker.port.addEventListener('message', this.step_func(function(e) {
    assert_true(e.data[0], "e.data === ''");
    assert_true(e.data[1], "e instanceof MessageEvent");
    assert_true(e.data[2], "e.ports.length == 1");
    this.done();
  }), false);
  worker.port.start();
});
</script>
